Remove old fragments for `VariedSubject` handling
