+++ app/src/main/java/com/senex/timetable/di/FragmentModule.kt
@@ -3,15 +3,11 @@ package com.senex.timetable.di
 import com.senex.timetable.presentation.ui.groups.GroupsFragment
 import com.senex.timetable.presentation.ui.schedule.ScheduleFragment
 import com.senex.timetable.presentation.ui.schedule.daily.DailyScheduleFragment
-import com.senex.timetable.presentation.ui.subject.elective.ElectiveSubjectFragment
-import com.senex.timetable.presentation.ui.subject.elective.selectable.SelectableElectiveSubjectsFragment
-import com.senex.timetable.presentation.ui.subject.english.EnglishSubjectFragment
-import com.senex.timetable.presentation.ui.subject.english.selectable.SelectableEnglishSubjectsFragment
 import com.senex.timetable.presentation.ui.subject.ordinary.OrdinarySubjectFragment
-import com.senex.timetable.presentation.ui.subject.varied.elective.NewElectiveSubjectFragment
-import com.senex.timetable.presentation.ui.subject.varied.elective.selectable.NewSelectableElectiveSubjectsFragment
-import com.senex.timetable.presentation.ui.subject.varied.english.NewEnglishSubjectFragment
-import com.senex.timetable.presentation.ui.subject.varied.english.selectable.NewSelectableEnglishSubjectsFragment
+import com.senex.timetable.presentation.ui.subject.varied.elective.ElectiveSubjectFragment
+import com.senex.timetable.presentation.ui.subject.varied.elective.selectable.SelectableElectiveSubjectsFragment
+import com.senex.timetable.presentation.ui.subject.varied.english.EnglishSubjectFragment
+import com.senex.timetable.presentation.ui.subject.varied.english.selectable.SelectableEnglishSubjectsFragment
 import dagger.Module
 import dagger.android.ContributesAndroidInjector
 
@@ -31,26 +27,14 @@ abstract class FragmentModule {
     abstract fun contributeOrdinarySubjectFragment(): OrdinarySubjectFragment
 
     @ContributesAndroidInjector
-    abstract fun contributeElectiveSubjectFragment(): ElectiveSubjectFragment
+    abstract fun contributeNewElectiveSubjectFragment(): ElectiveSubjectFragment
 
     @ContributesAndroidInjector
-    abstract fun contributeEnglishSubjectFragment(): EnglishSubjectFragment
+    abstract fun contributeNewEnglishSubjectFragment(): EnglishSubjectFragment
 
     @ContributesAndroidInjector
-    abstract fun contributeNewElectiveSubjectFragment(): NewElectiveSubjectFragment
+    abstract fun contributeNewSelectableElectiveSubjectsFragment(): SelectableElectiveSubjectsFragment
 
     @ContributesAndroidInjector
-    abstract fun contributeNewEnglishSubjectFragment(): NewEnglishSubjectFragment
-
-    @ContributesAndroidInjector
-    abstract fun contributeSelectableElectiveSubjectsFragment(): SelectableElectiveSubjectsFragment
-
-    @ContributesAndroidInjector
-    abstract fun contributeNewSelectableElectiveSubjectsFragment(): NewSelectableElectiveSubjectsFragment
-
-    @ContributesAndroidInjector
-    abstract fun contributeSelectableEnglishSubjectsFragment(): SelectableEnglishSubjectsFragment
-
-    @ContributesAndroidInjector
-    abstract fun contributeNewSelectableEnglishSubjectsFragment(): NewSelectableEnglishSubjectsFragment
+    abstract fun contributeNewSelectableEnglishSubjectsFragment(): SelectableEnglishSubjectsFragment
 }
