+++ app/src/main/java/com/senex/timetable/presentation/ui/subject/varied/elective/ElectiveSubjectViewModel.kt
@@ -5,18 +5,18 @@ import com.senex.timetable.domain.usecase.subject.GetAllByElectiveSubjectId
 import com.senex.timetable.domain.usecase.subject.elective.GetElectiveSubject
 import com.senex.timetable.domain.usecase.subject.elective.HideElectiveSubject
 import com.senex.timetable.domain.usecase.subject.elective.ShowElectiveSubject
-import com.senex.timetable.presentation.ui.subject.varied.BaseVariedSubjectViewModel
+import com.senex.timetable.presentation.ui.subject.varied.base.VariedSubjectViewModel
 import dagger.assisted.Assisted
 import dagger.assisted.AssistedFactory
 import dagger.assisted.AssistedInject
 
-class NewElectiveSubjectViewModel @AssistedInject constructor(
+class ElectiveSubjectViewModel @AssistedInject constructor(
     @Assisted electiveSubjectId: Long,
     showElectiveSubject: ShowElectiveSubject,
     hideElectiveSubject: HideElectiveSubject,
     getElectiveSubject: GetElectiveSubject,
     getAllByElectiveSubjectId: GetAllByElectiveSubjectId,
-) : BaseVariedSubjectViewModel<ElectiveSubject>(
+) : VariedSubjectViewModel<ElectiveSubject>(
     electiveSubjectId,
     showElectiveSubject,
     hideElectiveSubject,
@@ -25,6 +25,6 @@ class NewElectiveSubjectViewModel @AssistedInject constructor(
 ) {
     @AssistedFactory
     interface Factory {
-        fun create(electiveSubjectId: Long): NewElectiveSubjectViewModel
+        fun create(electiveSubjectId: Long): ElectiveSubjectViewModel
     }
 }
