+++ app/src/main/java/com/senex/timetable/presentation/ui/subject/varied/elective/selectable/SelectableElectiveSubjectsViewModel.kt
@@ -4,12 +4,12 @@ import com.senex.timetable.domain.model.subject.ElectiveSubject
 import com.senex.timetable.domain.usecase.subject.GetAllByElectiveSubjectId
 import com.senex.timetable.domain.usecase.subject.elective.GetElectiveSubject
 import com.senex.timetable.domain.usecase.subject.elective.SetPrimaryElectiveSubject
-import com.senex.timetable.presentation.ui.subject.varied.SelectableVariedSubjectsViewModel
+import com.senex.timetable.presentation.ui.subject.varied.base.selectable.SelectableVariedSubjectsViewModel
 import dagger.assisted.Assisted
 import dagger.assisted.AssistedFactory
 import dagger.assisted.AssistedInject
 
-class NewSelectableElectiveSubjectsViewModel @AssistedInject constructor(
+class SelectableElectiveSubjectsViewModel @AssistedInject constructor(
     @Assisted electiveSubjectId: Long,
     @Assisted savedPrimarySubjectId: Long?,
     setPrimaryElectiveSubject: SetPrimaryElectiveSubject,
@@ -27,6 +27,6 @@ class NewSelectableElectiveSubjectsViewModel @AssistedInject constructor(
         fun create(
             electiveSubjectId: Long,
             primarySubjectId: Long?,
-        ): NewSelectableElectiveSubjectsViewModel
+        ): SelectableElectiveSubjectsViewModel
     }
 }
