+++ app/src/main/java/com/senex/timetable/presentation/ui/subject/varied/english/selectable/SelectableEnglishSubjectsFragment.kt
@@ -3,16 +3,16 @@ package com.senex.timetable.presentation.ui.subject.varied.english.selectable
 import androidx.navigation.fragment.navArgs
 import com.senex.timetable.domain.model.subject.EnglishSubject
 import com.senex.timetable.presentation.common.assistedViewModel
-import com.senex.timetable.presentation.ui.subject.varied.SelectableVariedSubjectsFragment
+import com.senex.timetable.presentation.ui.subject.varied.base.selectable.SelectableVariedSubjectsFragment
 import javax.inject.Inject
 
-class NewSelectableEnglishSubjectsFragment :
+class SelectableEnglishSubjectsFragment :
     SelectableVariedSubjectsFragment<EnglishSubject>() {
-    private val args: NewSelectableEnglishSubjectsFragmentArgs by navArgs()
+    private val args: SelectableEnglishSubjectsFragmentArgs by navArgs()
 
     @Inject
-    lateinit var factory: NewSelectableEnglishSubjectsViewModel.Factory
-    override val viewModel: NewSelectableEnglishSubjectsViewModel by assistedViewModel {
+    lateinit var factory: SelectableEnglishSubjectsViewModel.Factory
+    override val viewModel: SelectableEnglishSubjectsViewModel by assistedViewModel {
         factory.create(
             args.englishSubjectId,
             args.primaryEnglishSubjectId.takeIf { it != -1L }
