+++ app/src/main/res/navigation/nav_graph_main.xml
@@ -16,17 +16,11 @@
             android:id="@+id/action_scheduleFragment_to_ordinarySubjectFragment"
             app:destination="@id/ordinarySubjectFragment" />
         <action
-            android:id="@+id/action_scheduleFragment_to_electiveSubjectFragment"
-            app:destination="@id/electiveSubjectFragment" />
-        <action
             android:id="@+id/action_scheduleFragment_to_settingsFragment"
             app:destination="@id/settingsFragment" />
         <action
             android:id="@+id/action_scheduleFragment_to_hiddenSubjectsFragment"
             app:destination="@id/hiddenSubjectsFragment" />
-        <action
-            android:id="@+id/action_scheduleFragment_to_englishSubjectFragment"
-            app:destination="@id/englishSubjectFragment" />
         <action
             android:id="@+id/action_scheduleFragment_to_newElectiveSubjectFragment"
             app:destination="@id/newElectiveSubjectFragment" />
@@ -56,66 +50,6 @@
             app:argType="long" />
     </fragment>
     <fragment
-        android:id="@+id/electiveSubjectFragment"
-        android:name="com.senex.timetable.presentation.ui.subject.elective.ElectiveSubjectFragment"
-        android:label="ElectiveSubjectFragment"
-        tools:layout="@layout/fragment_elective_subject">
-        <action
-            android:id="@+id/action_electiveSubjectFragment_to_scheduleFragment"
-            app:destination="@id/scheduleFragment" />
-        <argument
-            android:name="elective_subject_id"
-            app:argType="long" />
-        <action
-            android:id="@+id/action_electiveSubjectFragment_to_selectableElectiveSubjectsFragment"
-            app:destination="@id/selectableElectiveSubjectsFragment" />
-    </fragment>
-    <fragment
-        android:id="@+id/selectableElectiveSubjectsFragment"
-        android:name="com.senex.timetable.presentation.ui.subject.elective.selectable.SelectableElectiveSubjectsFragment"
-        android:label="SelectableElectiveSubjectsFragment"
-        tools:layout="@layout/fragment_selectable_elective_subjects">
-        <action
-            android:id="@+id/action_selectableElectiveSubjectsFragment_to_electiveSubjectFragment"
-            app:destination="@id/electiveSubjectFragment" />
-        <argument
-            android:name="elective_subject_id"
-            app:argType="long" />
-        <argument
-            android:name="primary_elective_subject_id"
-            app:argType="long" />
-    </fragment>
-    <fragment
-        android:id="@+id/englishSubjectFragment"
-        android:name="com.senex.timetable.presentation.ui.subject.english.EnglishSubjectFragment"
-        android:label="EnglishSubjectFragment"
-        tools:layout="@layout/fragment_english_subject">
-        <action
-            android:id="@+id/action_englishSubjectFragment_to_scheduleFragment"
-            app:destination="@id/scheduleFragment" />
-        <argument
-            android:name="english_subject_id"
-            app:argType="long" />
-        <action
-            android:id="@+id/action_englishSubjectFragment_to_selectableEnglishSubjectsFragment"
-            app:destination="@id/selectableEnglishSubjectsFragment" />
-    </fragment>
-    <fragment
-        android:id="@+id/selectableEnglishSubjectsFragment"
-        android:name="com.senex.timetable.presentation.ui.subject.english.selectable.SelectableEnglishSubjectsFragment"
-        android:label="SelectableEnglishSubjectsFragment"
-        tools:layout="@layout/fragment_selectable_english_subjects">
-        <action
-            android:id="@+id/action_selectableEnglishSubjectsFragment_to_englishSubjectFragment"
-            app:destination="@id/englishSubjectFragment" />
-        <argument
-            android:name="english_subject_id"
-            app:argType="long" />
-        <argument
-            android:name="primary_english_subject_id"
-            app:argType="long" />
-    </fragment>
-    <fragment
         android:id="@+id/settingsFragment"
         android:name="com.senex.timetable.presentation.ui.settings.SettingsFragment"
         android:label="SettingsFragment"
@@ -135,7 +69,7 @@
     </fragment>
     <fragment
         android:id="@+id/newElectiveSubjectFragment"
-        android:name="com.senex.timetable.presentation.ui.subject.varied.elective.NewElectiveSubjectFragment"
+        android:name="com.senex.timetable.presentation.ui.subject.varied.elective.ElectiveSubjectFragment"
         android:label="NewElectiveSubjectFragment"
         tools:layout="@layout/fragment_varied_subject">
         <action
@@ -150,7 +84,7 @@
     </fragment>
     <fragment
         android:id="@+id/newEnglishSubjectFragment"
-        android:name="com.senex.timetable.presentation.ui.subject.varied.english.NewEnglishSubjectFragment"
+        android:name="com.senex.timetable.presentation.ui.subject.varied.english.EnglishSubjectFragment"
         android:label="NewEnglishSubjectFragment"
         tools:layout="@layout/fragment_varied_subject">
         <action
@@ -165,7 +99,7 @@
     </fragment>
     <fragment
         android:id="@+id/newSelectableElectiveSubjectsFragment"
-        android:name="com.senex.timetable.presentation.ui.subject.varied.elective.selectable.NewSelectableElectiveSubjectsFragment"
+        android:name="com.senex.timetable.presentation.ui.subject.varied.elective.selectable.SelectableElectiveSubjectsFragment"
         android:label="NewSelectableElectiveSubjectsFragment"
         tools:layout="@layout/fragment_selectable_varied_subjects">
         <action
@@ -180,7 +114,7 @@
     </fragment>
     <fragment
         android:id="@+id/newSelectableEnglishSubjectsFragment"
-        android:name="com.senex.timetable.presentation.ui.subject.varied.english.selectable.NewSelectableEnglishSubjectsFragment"
+        android:name="com.senex.timetable.presentation.ui.subject.varied.english.selectable.SelectableEnglishSubjectsFragment"
         android:label="NewSelectableEnglishSubjectsFragment"
         tools:layout="@layout/fragment_selectable_varied_subjects">
         <action
